Initial implementation of Chalvatzis architecture
This version is an initial implementation but does not have the full architecture yet as it is a many to one vs many to many that was used in the paper. However, it is learning (and quite likely overfitting) with the same data set that Chalvatzis used in his paper.