Changed to seq to seq model
Expanded training label to be the sequence of adjusted close prices up to time t plus the future price t+1.  Changed loss function to be computing the difference between the  complete output sequence from the LSTM (vs just using the last hidden state as in all previous versions) and the new label.